+++ Core/Abstractions/IBookService.cs
@@ -8,6 +8,6 @@ namespace Application
         Task<Guid> DeleteBook(Guid id);
         Task<List<Book>> GetAllBooks();
         Task<Book> GetBook(Guid id);
-        Task<Guid> UpdateBook(Guid id, string title, string description, int price);
+        Task<Guid> UpdateBook(Guid id, string title, string description, decimal price);
     }
 }
